refactor: main.js 코드 분리 및 아몬드 컬렉션 상품 누락 버그 수정
- 메인 페이지 전용 로직을 src/pages/main/index.js로 분리
- main.js에는 전역 공통 코드(CSS, 토큰 검증)만 유지
- fetchCategories 불필요 로직 제거 후 limit:100으로 전체 상품 조회
- if/else 구문 중괄호 누락 린트 에러 수정
